# Refactor State to utilize SimpleNamespace for improved perfomance