--- a/starlette/datastructures.py
+++ b/starlette/datastructures.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import typing
 from shlex import shlex
+from types import SimpleNamespace
 from urllib.parse import SplitResult, parse_qsl, urlencode, urlsplit
 
 from starlette.concurrency import run_in_threadpool
@@ -677,7 +678,7 @@ class MutableHeaders(Headers):
         self["vary"] = vary
 
 
-class State:
+class State(SimpleNamespace):
     """
     An object that can be used to store arbitrary state.
 
@@ -690,16 +691,8 @@ class State:
         if state is None:
             state = {}
         super().__setattr__("_state", state)
+        super().__init__(**state)
 
     def __setattr__(self, key: typing.Any, value: typing.Any) -> None:
+        super().__setattr__(key, value)
         self._state[key] = value
-
-    def __getattr__(self, key: typing.Any) -> typing.Any:
-        try:
-            return self._state[key]
-        except KeyError:
-            message = "'{}' object has no attribute '{}'"
-            raise AttributeError(message.format(self.__class__.__name__, key))
-
-    def __delattr__(self, key: typing.Any) -> None:
-        del self._state[key]