--- a/starlette/datastructures.py
+++ b/starlette/datastructures.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import typing
 from shlex import shlex
-from types import SimpleNamespace
 from urllib.parse import SplitResult, parse_qsl, urlencode, urlsplit
 
 from starlette.concurrency import run_in_threadpool
@@ -678,7 +677,7 @@ class MutableHeaders(Headers):
         self["vary"] = vary
 
 
-class State(SimpleNamespace):
+class State:
     """
     An object that can be used to store arbitrary state.
 
@@ -691,8 +690,17 @@ class State(SimpleNamespace):
         if state is None:
             state = {}
         super().__setattr__("_state", state)
-        super().__init__(**state)
 
     def __setattr__(self, key: typing.Any, value: typing.Any) -> None:
-        super().__setattr__(key, value)
         self._state[key] = value
+
+    def __getattribute__(self, key: typing.Any) -> typing.Any:
+        state = object.__getattribute__(self, "_state")
+
+        if key in state:
+            return state[key]
+
+        return object.__getattribute__(self, key)
+
+    def __delattr__(self, key: typing.Any) -> None:
+        del self._state[key]