--- a/starlette/datastructures.py
+++ b/starlette/datastructures.py
@@ -692,7 +692,7 @@ class State:
         super().__setattr__("_state", state)
 
     def __setattr__(self, key: typing.Any, value: typing.Any) -> None:
-        self._state[key] = value
+        object.__getattribute__(self, "_state")[key] = value
 
     def __getattribute__(self, key: typing.Any) -> typing.Any:
         state = object.__getattribute__(self, "_state")