@Zaczero 2024-04-03
Context: I am just a random person - not a maintainer. I personally think the SimpleNamespace would work best but it can be simplified like that:

```py
class NameSpaceState2(SimpleNamespace):
    """
    An object that can be used to store arbitrary state.

    Used for `request.state` and `app.state`.
    """

    def __init__(self, state: dict[str, typing.Any] | None = None):
        if state is not None:
            super().__init__(**state)

    @property
    def _state(self):
        return self.__dict__
```

There is no need to keep 2 separate dicts. It has better performance characteristic for setting attrs and does not bug out when accessing _state directly.

![image](https://github.com/encode/starlette/assets/10835147/c40c9a78-0099-41d5-9af3-2e026fa3b948)

![image](https://github.com/encode/starlette/assets/10835147/1afa0a1c-8160-4e8a-866a-196d83ad4c8a)

I confirm that this class is generally 10x faster compared to existing solution.