@Kludex 2024-12-12
> Context: I am just a random person - not a maintainer. I personally think the SimpleNamespace would work best but it can be simplified like that:
> 
> ```python
> class NameSpaceState2(SimpleNamespace):
>     """
>     An object that can be used to store arbitrary state.
> 
>     Used for `request.state` and `app.state`.
>     """
> 
>     def __init__(self, state: dict[str, typing.Any] | None = None):
>         if state is not None:
>             super().__init__(**state)
> 
>     @property
>     def _state(self):
>         return self.__dict__
> ```
> 
> There is no need to keep 2 separate dicts. It has better performance characteristic for setting attrs and does not bug out when accessing _state directly.
> 
> ![image](https://private-user-images.githubusercontent.com/10835147/319344267-c40c9a78-0099-41d5-9af3-2e026fa3b948.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MzQwMTAwMDIsIm5iZiI6MTczNDAwOTcwMiwicGF0aCI6Ii8xMDgzNTE0Ny8zMTkzNDQyNjctYzQwYzlhNzgtMDA5OS00MWQ1LTlhZjMtMmUwMjZmYTNiOTQ4LnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNDEyMTIlMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjQxMjEyVDEzMjE0MlomWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPWRjNzllOTgxY2YyYWU3ZGQwNTE5MjZhYzYzMTI4YzAwMTJhY2JjODdiN2FmMGE0ZDdlZTI3MWVjMDA0MWJlNDkmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0In0.7_nqcdmZZ7cGHzn4awiNzSmniCtUiyjuihkfjO7HpeY)
> 
> ![image](https://private-user-images.githubusercontent.com/10835147/319344378-1afa0a1c-8160-4e8a-866a-196d83ad4c8a.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MzQwMTAwMDIsIm5iZiI6MTczNDAwOTcwMiwicGF0aCI6Ii8xMDgzNTE0Ny8zMTkzNDQzNzgtMWFmYTBhMWMtODE2MC00ZThhLTg2NmEtMTk2ZDgzYWQ0YzhhLnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNDEyMTIlMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjQxMjEyVDEzMjE0MlomWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPTllMmE5ZjIxYmE0ZGRhMWMzYzg2OTkxYmRkN2ViMjBhY2JiN2ZhODE4YWM1YTA1MTNlMDNkOTc3MmJmOTE5MTUmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0In0.Mv8a60TAsg861zA0EhqTNHvEZycLsvfu3-RkWot0ed4)
> 
> I confirm that this class is generally 10x faster compared to existing solution.

This fails on the test suite 🤔